% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{AddAResponse}
\alias{AddAResponse}
\title{AddAResponse}
\usage{
AddAResponse(RandoDataFrame, nobs = NA, ResponseUnit = "", RespValues)
}
\arguments{
\item{RandoDataFrame}{a TrainRandoSubj object with the randomization data}

\item{nobs}{the number of observations in the dataset by default
the number of rows in RandoDataFrame}

\item{ResponseUnit}{a character string with the unit of the response}

\item{RespValues}{can be a fixed value intercept which will be repeated nobs times
or a numerical vector of the same lenght than nobs.}
}
\value{
response TrainRandoResp object with a response variable
 (dependent variable)
}
\description{
AddAResponse
}
\examples{
data("myArms","myRandoDataFrame","myprogfact")
myResponse <- AddAResponse(RandoDataFrame=myRandoDataFrame, ResponseUnit='Weight (in kg)',RespValues=70)
}
