% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddNewSubject.R
\name{AddFirstSubject}
\alias{AddFirstSubject}
\title{AddFirstSubject}
\usage{
AddFirstSubject(progfact, Arms, usubjid = "Pat-0001")
}
\arguments{
\item{progfact}{A TrainRandoProgFact object defining the prognostic factors}

\item{Arms}{A TrainRandoArm object defining the treatment arms}
}
\value{
RandoDataFrame TrainRandoSubj object witht the randomization data
}
\description{
AddFirstSubject
}
\examples{
data("myArms","myprogfact")
myRandoDataFrame <- AddFirstSubject(progfact=trainingrando::myprogfact,Arms=trainingrando::myArms)
}
