% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrgFact.R
\name{AddNewPrgFact}
\alias{AddNewPrgFact}
\title{Add a new training factor in an existing TrainRandoProgFact object.}
\usage{
AddNewPrgFact(
  currentProgfact,
  progfactLabel = "Prognostic factor 2",
  progfactLevels = c("A", "B", "C")
)
}
\arguments{
\item{currentProgfact}{Une liste de classe SAPUL4012Moy2EchPair.}

\item{progfactLabel}{a character string with the name of the prognostic factor.}

\item{progfactLevels}{a vector of character strings with the levels of the prognostic factor.}
}
\value{
NewProgfact Une liste de classe SAPUL4012Moy2EchPair.
}
\description{
This function adds a new prognostic factor in an existing TrainRandoProgFact object.
}
\examples{
myprogfact2 <- AddNewPrgFact(currentProgfact=myprogfact,progfactLabel = "Type of tumor",progfactLevels=c("A", "B","C"))
}
\seealso{
CreatePrgFact
}
