% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddNewSubject.R
\name{AddNewSubject}
\alias{AddNewSubject}
\title{AddNewSubject.}
\usage{
AddNewSubject(progfact, usubjid)
}
\arguments{
\item{progfact}{A TrainRandoProgFact object defining the prognostic factors}

\item{usubjid}{the identifier of the subject (that is the id of the patient)}
}
\value{
NewSubj , A new patient with the prognostic level randomly assigned according to
according to independant uniform distribution.
}
\description{
Add a new subject based on the prognostic factor and the id of the subject.
  This function is often used inside the function AddNextSubject.
}
\examples{
myNewSubj <- AddNewSubject(progfact=myprogfact,usubjid='Pat001')
}
