% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddNewSubject.R
\name{AddNextSubject}
\alias{AddNextSubject}
\title{AddNextSubject}
\usage{
AddNextSubject(RandoDataFrame, progfact, ...)
}
\arguments{
\item{ArmLevel}{A dataframe with ArmLevel$Name = c('TreatmentA','TreatmentB')}
}
\value{
ArmLevel a dataframe with LowerBound and Upperbound for the random allocation
}
\description{
AddNextSubject
}
\examples{
data("myArms","myRandoDataFrame","myprogfact")
myRandoDataFrame <- AddNextSubject(RandoDataFrame=trainingrando::myRandoDataFrame, progfact=trainingrando::myprogfact)
}
