% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmLevel.R
\name{ArmLevelProg}
\alias{ArmLevelProg}
\title{ArmLevelProg.}
\usage{
ArmLevelProg(ArmLevel)
}
\arguments{
\item{ArmLevel}{A dataframe with ArmLevel$Name = c('TreatmentA','TreatmentB')}
}
\value{
ArmLevel a dataframe with LowerBound and Upperbound for the random allocation
}
\description{
ArmLevelProg.
}
\examples{
myArms <- ArmLevelProg(data.frame("Name" = c('Placebo', 'Treat1','Treat2','Treat3')))
}
