% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrBetweenIndVar.R
\name{CorrBetweenIndVar}
\alias{CorrBetweenIndVar}
\title{CorrBetweenIndVar}
\usage{
CorrBetweenIndVar(RandoDataFrame)
}
\arguments{
\item{RandoDataFrame}{a TrainRandoSubj object witht the randomization data}
}
\value{
CorrVector , a vector with the correlation between the treatment arm and the other independent variables.
}
\description{
CorrBetweenIndVar
}
\examples{
CorrVector <- CorrBetweenIndVar(RandoDataFrame=myRandoDataFrame)
}
