% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrgFact.R
\name{CreatePrgFact}
\alias{CreatePrgFact}
\title{Create a new TrainRandoProgFact object.}
\usage{
CreatePrgFact(
  progfactLabel = "Prognostic factor 1",
  progfactLevels = c("A", "B", "C")
)
}
\arguments{
\item{progfactLabel}{a character string with the name of the prognostic factor.}

\item{progfactLevels}{a vector of character strings with the levels of the prognostic factor.}
}
\value{
Une liste de classe SAPUL4012Moy2EchPair.
}
\description{
This create a new list with a first prognostic factor
}
\examples{
myprogfact <- CreatePrgFact(progfactLabel = "Age group",progfactLevels=c("0-18 years", "19-35 years","36-60 years","> 60 years"))
}
\seealso{
AddNewPrgFact
}
