% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DicotomizeVector.R
\name{DicotomizeVector}
\alias{DicotomizeVector}
\title{DicotomizeVector}
\usage{
DicotomizeVector(v1)
}
\arguments{
\item{v1}{a vector of c factor levels}
}
\value{
a n x c boolean matrix
}
\description{
This create a n x c boolean matrix from a vector of c factor levels
}
