% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenStudyTrial.R
\name{GenStudyTrial}
\alias{GenStudyTrial}
\title{GenStudyTrial}
\usage{
GenStudyTrial(npat = 25, progfact, Arms)
}
\arguments{
\item{npat}{the number of patients that will be simulated in the trials}

\item{progfact}{A TrainRandoProgFact object defining the prognostic factors}

\item{Arms}{A TrainRandoArm object defining the treatment arms}
}
\value{
RandoDataFrame TrainRandoSubj object witht the randomization data
}
\description{
GenStudyTrial
}
\examples{
data("myArms","myprogfact")
RandoDataFrame <- GenStudyTrial(npat = 25, progfact = myprogfact, Arms=myArms)
}
