% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myArms}
\alias{myArms}
\title{Object specifying the arms of a study.}
\format{A data frame with 4 rows and 5 variables:
\describe{
  \item{Name}{The name of the treatment arms (Factor)}
  \item{Prob}{The probability of each treatment arms (num)}
  \item{CumProb}{The cumulative probability (num)}
  \item{LowerBound}{ (num)}
  \item{UpperBound}{ (num)}
}}
\usage{
myArms
}
\description{
A a dataframe and TrainRandoArm with LowerBound and Upperbound for the random allocation.
}
\keyword{datasets}
