% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myepsilon}
\alias{myepsilon}
\title{Epsilon object with the random errors and the two keys id variables (usubjidd, TimePoint).}
\format{A list with the following elements:
\describe{
\item{usubjid}{the unique subject id (char)}
\item{TimePoint}{the time points (char)}
\item{Residuals}{the residuals i.e random error (num)}
}}
\usage{
myepsilon
}
\description{
A data.frame, tbl and tbl_df object defining with the random error and the keys id variables.
}
\keyword{datasets}
