% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myprogfact}
\alias{myprogfact}
\title{Object specifying the prognostic factor of a disease.}
\format{A list in which each element corresponds to a prognostic factor.
\describe{
}}
\usage{
myprogfact
}
\description{
A a list and TrainRandoProgFact object defining the progostic factor of a disease.
}
\keyword{datasets}
