% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myRandoDataFrame}
\alias{myRandoDataFrame}
\title{Object TrainRandoSubj object with the randomization data.}
\format{A list in which each element corresponds to a prognostic factor.
\describe{
\item{usubjid}{the unique subject id (char)}
\item{V2}{the age group (num with factor)}
\item{V3}{the type of tumor (num with factor)}
\item{V4}{the genotype (num with factor)}
\item{Arm}{the arms (factor)}
}}
\usage{
myRandoDataFrame
}
\description{
A a data.frame and TrainRandoSubj object defining the progostic factor of a disease.
}
\keyword{datasets}
