% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myResponse}
\alias{myResponse}
\title{Object TrainRandoResp object with the response (dependent variable).}
\format{A list with the following elements:
\describe{
\item{usubjid}{the unique subject id (char)}
\item{V2}{the age group (num with factor)}
\item{V3}{the type of tumor (num with factor)}
\item{V4}{the genotype (num with factor)}
\item{Arm}{the arms (factor)}
}}
\usage{
myResponse
}
\description{
A list and TrainRandoResp object defining the progostic factor of a disease.
}
\keyword{datasets}
