% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myVarCovMatr}
\alias{myVarCovMatr}
\title{A variance covariance matrix.}
\format{A square positive defined matrix
\describe{
}}
\usage{
myVarCovMatr
}
\description{
A variance covariance matrix for a linear model with repeated measures.
}
\keyword{datasets}
