% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddNewSubject.R
\name{print.TrainRandoSubj}
\alias{print.TrainRandoSubj}
\title{Print a TrainRandoSubj object}
\usage{
\method{print}{TrainRandoSubj}(x, ...)
}
\arguments{
\item{x}{Un a object TrainRandoSubj}
}
\description{
Print a TrainRandoSubj object
}
