% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurlyRandom.R
\name{PurlyRandom}
\alias{PurlyRandom}
\title{PurlyRandom}
\usage{
PurlyRandom(Arms, RandoDataFrame, usubjid = NULL)
}
\arguments{
\item{Arms}{A list with the level of the prognostic factors}

\item{RandoDataFrame}{a TrainRandoSubj object witht the randomization data}

\item{usubjid}{the identifier of the subject (that is the id of the patient)}
}
\value{
RandoDataFrame , a TrainRandoSubj object witht the randomization
 data where the patient given in usubjid has a new treatment allocation
according to independant uniform distribution.
}
\description{
PurlyRandom
}
\examples{
myRandoDataFrame02 <- PurlyRandom(Arms=myArms,RandoDataFrame=myRandoDataFrame,usubjid="Pat-0002")
}
