% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{RespResidVar}
\alias{RespResidVar}
\title{RespResidVar}
\usage{
RespResidVar(Response, SDReplicates)
}
\arguments{
\item{Response}{TrainRandoResp object with a response variable (dependent variable)}

\item{SDReplicates}{a character string with the unit of the response}
}
\value{
response TrainRandoResp object with a response variable (dependent variable)
}
\description{
RespResidVar
}
\examples{
data("myResponse")
myResponse <- RespResidVar(Response=myResponse,SDReplicates=10)
}
