% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovarianceStructure.R
\name{UniformCorrelation}
\alias{UniformCorrelation}
\title{UniformCorrelation}
\usage{
UniformCorrelation(rho, sigma2, nTimePoints)
}
\arguments{
\item{rho}{the pearson correlation parameter (should be between -1 and 1)}

\item{sigma2}{the residual variance}

\item{nTimePoints}{the number of timepoints}
}
\value{
VarCovMatr a VarCovMatr object which is a squared matrix defining the covariance structure.
}
\description{
UniformCorrelation
}
\examples{
VarCovMatr <- UniformCorrelation(rho=0.5,sigma2=2,nTimePoints=6)
}
