% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{AddAResponse}
\alias{AddAResponse}
\title{AddAResponse}
\usage{
AddAResponse(RandoDataFrame, ResponseUnit = "", Intercept)
}
\arguments{
\item{RandoDataFrame}{a TrainRandoSubj object witht the randomization data}

\item{ResponseUnit}{a character string with the unit of the response}

\item{Intercept}{a fixed value}
}
\value{
response TrainRandoResp object witht a response variable (dependent variable)
}
\description{
AddAResponse
}
\examples{
data("myArms","myRandoDataFrame","myprogfact")
myResponse <- AddAResponse(RandoDataFrame=myRandoDataFrame,ResponseUnit='Weight (in kg)',Intercept=70)
}
