% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddNewSubject.R
\name{AddNewSubject}
\alias{AddNewSubject}
\title{AddNewSubject.}
\usage{
AddNewSubject(lsPrgFact, usubjid)
}
\arguments{
\item{lsPrgFact}{A list with the level of the prognostic factors}

\item{usubjid}{the identifier of the subject (that is the id of the patient)}
}
\value{
NewSubj , A new patient with the prognostic level randomly assigned according to
according to independant uniform distribution.
}
\description{
AddNewSubject.
}
\examples{
progfact1<-c("A","B","C")
progfact2<-c("A","B","C")
progfact3<-c("A","B","C","D")
PrgFact<-list(progfact1, progfact2, progfact3)
myNewSubj <- AddNewSubject(lsPrgFact=myprogfact,usubjid='Pat001')
}
