% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddNewSubject.R
\name{AddNextSubject}
\alias{AddNextSubject}
\title{AddNextSubject}
\usage{
AddNextSubject(RandoDataFrame, progfact)
}
\arguments{
\item{RandoDataFrame}{a TrainRandoSubj object witht the randomization data}

\item{progfact}{A TrainRandoProgFact object defining the prognostic factors}
}
\value{
RandoDataFrame a TrainRandoSubj object witht the randomization data
}
\description{
AddNextSubject
}
\examples{
data("myArms","myRandoDataFrame","myprogfact")
myRandoDataFrame <- AddNextSubject(RandoDataFrame=trainingrando::myRandoDataFrame, progfact=trainingrando::myprogfact)
}
