% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmLevel.R
\name{print.TrainRandoArm}
\alias{print.TrainRandoArm}
\title{print a TrainRandoArm object}
\usage{
\method{print}{TrainRandoArm}(x, ...)
}
\arguments{
\item{x}{Un a object TrainRandoArm}
}
\description{
print a TrainRandoArm object
}
