% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrgFact.R
\name{print.TrainRandoProgFact}
\alias{print.TrainRandoProgFact}
\title{Print a TrainRandoProgFact object}
\usage{
\method{print}{TrainRandoProgFact}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a TrainRandoProgFact object
}
