% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovarianceStructure.R
\name{RandomEffect}
\alias{RandomEffect}
\title{RandomEffect}
\usage{
RandomEffect(randomeffect, VARRandoEffect)
}
\arguments{
\item{randomeffect}{an independent variable which is modeled as a random effect.}

\item{VARRandoEffect}{The variance of the random effect}
}
\value{
ZGamma1
}
\description{
Creation of a vector of the random effect
}
\examples{
myZGamma <- RandomEffect(randomeffect=myRandoDataFrame$V2, VARRandoEffect=10)
}
