% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{RespAddFixedEffect}
\alias{RespAddFixedEffect}
\title{RespAddFixedEffect}
\usage{
RespAddFixedEffect(Response, RandoDataFrame, ColID, BetaVector)
}
\arguments{
\item{Response}{TrainRandoResp object with a response variable (dependent variable)}

\item{RandoDataFrame}{a TrainRandoSubj object with the randomization data}

\item{ColID}{index of the column of the RandoDataFrame which will be used as independant variable}

\item{BetaVector}{a vector of the beta effect}
}
\value{
response TrainRandoResp object with a response variable (dependent variable)
}
\description{
RespAddFixedEffect
}
\examples{
data("myResponse","myRandoDataFrame")
myResponse <- RespAddFixedEffect(Response=trainingrando::myResponse,RandoDataFrame=trainingrando::myRandoDataFrame,ColID=2, BetaVector=c(-1,0,0,1))
}
